Extract featureMock to dedicated file
